+++ src/index.html
@@ -23,6 +23,6 @@
     <!-- <include src="./partials/burger-menu.html"></include> -->
     <!-- Scripts -->
     <script src="./js/categories.js"></script>
-    <script type="module" src="./js/axios/axios-weather.js"></script>
+    <script type="module" src="./js/requests/weatherFetch.js"></script>
   </body>
 </html>
